+++ source/clSliceData.h
@@ -87,7 +87,7 @@ public:
 	/// <summary>returnes the number of coordinate points of the part object [objectIndex]</summary>
 	/// <param name="partIndex">index of the Part</param>
 	/// <param name="objectIndex">index of the polyline or Hatch of the part</param>
-	float getObjectPointCount(int partIndex, int objectIndex);
+	int getObjectPointCount(int partIndex, int objectIndex);
 
 	/// <summary>returnes the coordinates of the part object [objectIndex]</summary>
 	/// <param name="partIndex">index of the Part</param>
